+++ src/components/pages/apps/genes/gene-convert/gene-convert-page.tsx
@@ -1,12 +1,9 @@
 'use client'
 
-import { ToolbarOpenFile } from '@/toolbar/toolbar-open-files'
-
 import { TabbedDataFrames } from '@/components/table/tabbed-dataframes'
 
 import { FooterPortal } from '@/components/toolbar/footer-portal'
 
-import { PlayIcon } from '@/icons/play-icon'
 import {
   ShowOptionsMenu,
   Toolbar,
@@ -14,20 +11,9 @@ import {
   ToolbarPanel,
 } from '@/toolbar/toolbar'
 
-import {
-  downloadDataFrame,
-  getFormattedShape,
-} from '@/lib/dataframe/dataframe-utils'
+import { getFormattedShape } from '@/lib/dataframe/dataframe-utils'
 
-import {
-  DEFAULT_PARSE_OPTS,
-  filesToDataFrames,
-  onTextFileChange,
-  type IParseOptions,
-  type ITextFileOpen,
-} from '@/components/pages/open-files'
-
-import { ToolbarTabGroup } from '@/toolbar/toolbar-tab-group'
+import { onTextFileChange } from '@/components/pages/open-files'
 
 import { OpenIcon } from '@/icons/open-icon'
 
@@ -38,26 +24,22 @@ import {
   TEXT_DOWNLOAD_AS_TXT,
   TEXT_OPEN_FILE,
   TEXT_SAVE_AS,
-  TEXT_SAVE_TABLE,
 } from '@/consts'
 
 import { DropdownMenuItem } from '@/components/shadcn/ui/themed/v2/dropdown-menu'
 import { UploadIcon } from '@/icons/upload-icon'
-import { createGeneConvTable, type Species } from '@/lib/gene/geneconv'
 
 import { ShortcutLayout } from '@/layouts/shortcut-layout'
 
 import type { ITab } from '@/components/tabs/tab-provider'
 import { FileIcon } from '@/icons/file-icon'
 import type { AnnotationDataFrame } from '@/lib/dataframe/annotation-dataframe'
-import { ToolbarIconButton } from '@/toolbar/toolbar-icon-button'
 import { ZoomSlider } from '@/toolbar/zoom-slider'
 
 import { useDialogs } from '@/components/dialogs/dialogs'
 import { AppHeaderIcon } from '@/components/header/app-header-icon'
 import { AppInfoButton } from '@/components/header/app-info-button'
 import { HeaderSlotPortal } from '@/components/header/header-portal'
-import { DownloadIcon } from '@/components/icons/download-icon'
 import { useStableId } from '@/hooks/stable-id'
 import { HeaderButton } from '@/layouts/header-button'
 import { DataFrameReader } from '@/lib/dataframe/dataframe-reader'
@@ -66,19 +48,22 @@ import { useAppInfo } from '@/lib/edb/edb-settings'
 import { httpFetch } from '@/lib/http/http-fetch'
 import { textToLines } from '@/lib/text/lines'
 import { CoreProviders } from '@/providers/core-provider'
-import { GroupToggle, ToggleGroup } from '@/themed/v2/toggle-group'
 import {
   HistoryLayout,
   HistoryShowButton,
 } from '../../matcalc/history/history-layout'
 
+import { useToolbarTabs } from '@/components/tabs/tab-store'
 import {
   useCurrentSheets,
   useFiles,
 } from '../../matcalc/history/history-provider/history-contexts'
 import { useHistory } from '../../matcalc/history/history-provider/history-provider'
 import { UndoShortcuts } from '../../matcalc/history/undo-shortcuts'
+import { useOpenFiles } from '../../matcalc/hooks/open'
+import { useSave } from '../../matcalc/hooks/save'
 import APP_INFO from './manifest.json'
+import { HomeToolbar } from './toolbars/home-toolbar'
 
 export function GeneConvPage() {
   const _id = useStableId('gene-convert-page')
@@ -90,74 +75,37 @@ export function GeneConvPage() {
 
   const { setAppInfo } = useAppInfo()
 
-  const [fromSpecies, setFromSpecies] = useState<Species>('human')
-  const [toSpecies, setToSpecies] = useState<Species>('mouse')
-  const [exact] = useState(true)
-
   const { open: openDialog } = useDialogs()
   const [showSideBar, setShowSideBar] = useState(true)
 
   const [showFileMenu, setShowFileMenu] = useState(false)
 
-  const speciesTabs = [
-    { id: 'human', name: 'Human' },
-    { id: 'mouse', name: 'Mouse' },
-  ]
+  const { setTabs: setToolbarTabs } = useToolbarTabs()
+
+  const { openFiles } = useOpenFiles()
+  const { save } = useSave()
 
   useEffect(() => {
     setAppInfo(APP_INFO)
-  }, [])
+  }, [setAppInfo])
 
-  function openFiles(
-    files: ITextFileOpen[],
-    options: IParseOptions = DEFAULT_PARSE_OPTS
-  ) {
-    //filesToDataFrames(files, historyDispatch, options)
+  // function openFiles(
+  //   files: ITextFileOpen[],
+  //   options: IParseOptions = DEFAULT_PARSE_OPTS
+  // ) {
+  //   //filesToDataFrames(files, historyDispatch, options)
 
-    filesToDataFrames(files, {
-      parseOpts: options,
-      onSuccess: (tables) => {
-        if (tables.length > 0) {
-          openFile(tables[0]!.name, { sheets: tables })
-        }
-      },
-    })
+  //   filesToDataFrames(files, {
+  //     parseOpts: options,
+  //     onSuccess: (tables) => {
+  //       if (tables.length > 0) {
+  //         openFile(tables[0]!.name, { sheets: tables })
+  //       }
+  //     },
+  //   })
 
-    setShowFileMenu(false)
-  }
-
-  async function convertGenes() {
-    if (!sheet) {
-      return
-    }
-
-    const dfa = await createGeneConvTable(sheet as AnnotationDataFrame, {
-      fromSpecies,
-      toSpecies,
-      exact,
-    })
-
-    if (dfa) {
-      addSheets([dfa], { name: `Gene Conversion` })
-    }
-  }
-
-  function save(format: string) {
-    if (!sheet) {
-      return
-    }
-
-    const sep = format === 'csv' ? ',' : '\t'
-
-    downloadDataFrame(sheet as AnnotationDataFrame, {
-      hasHeader: true,
-      hasIndex: false,
-      file: `table.${format}`,
-      sep,
-    })
-
-    setShowFileMenu(false)
-  }
+  //   setShowFileMenu(false)
+  // }
 
   // Load a default sheet
   // useEffect(() => {
@@ -179,96 +127,6 @@ export function GeneConvPage() {
       // do nothing
     }
   }
-
-  const tabs: ITab[] = [
-    {
-      //id: nanoid(),
-      id: 'Home',
-      component: () => (
-        <>
-          <ToolbarTabGroup title="File">
-            <ToolbarOpenFile
-              onOpen={() => {
-                openDialog({
-                  type: 'open',
-                  payload: {
-                    callback: (message, files) => {
-                      onTextFileChange(message, files, (files) =>
-                        openFiles(files)
-                      )
-                    },
-                  },
-                })
-              }}
-              multiple={true}
-            />
-
-            <ToolbarIconButton
-              onClick={() => {
-                openDialog({
-                  type: 'save',
-                  payload: {
-                    callback: (data) => {
-                      save(data.format.ext)
-                    },
-                  },
-                })
-              }}
-              title={TEXT_SAVE_TABLE}
-            >
-              <DownloadIcon />
-            </ToolbarIconButton>
-          </ToolbarTabGroup>
-
-          <ToolbarTabGroup title="Convert">
-            <ToolbarIconButton
-              aria-label="Convert"
-              onClick={convertGenes}
-              title="Convert"
-            >
-              <PlayIcon />
-            </ToolbarIconButton>
-          </ToolbarTabGroup>
-
-          <ToolbarTabGroup className="gap-x-2 mr-1" title="From">
-            <ToggleGroup
-              //variant="outline"
-
-              value={[fromSpecies]}
-              onValueChange={(v) => {
-                setFromSpecies(v[0]! as Species)
-              }}
-              size="toolbar"
-              direction="toolbar"
-            >
-              {speciesTabs.map((tab) => (
-                <GroupToggle key={tab.id} value={tab.id}>
-                  {tab.name}
-                </GroupToggle>
-              ))}
-            </ToggleGroup>
-          </ToolbarTabGroup>
-
-          <ToolbarTabGroup className="gap-x-2 ml-1" title="To">
-            <ToggleGroup
-              value={[toSpecies]}
-              onValueChange={(v) => {
-                setToSpecies(v[0]! as Species)
-              }}
-              size="toolbar"
-              direction="toolbar"
-            >
-              {speciesTabs.map((tab) => (
-                <GroupToggle key={tab.id} value={tab.id}>
-                  {tab.name}
-                </GroupToggle>
-              ))}
-            </ToggleGroup>
-          </ToolbarTabGroup>
-        </>
-      ),
-    },
-  ]
 
   const fileMenuTabs: ITab[] = [
     {
@@ -317,6 +175,18 @@ export function GeneConvPage() {
     },
   ]
 
+  useEffect(() => {
+    const tabs: ITab[] = [
+      {
+        //id: nanoid(),
+        id: 'Home',
+        component: HomeToolbar,
+      },
+    ]
+
+    setToolbarTabs(tabs)
+  }, [setToolbarTabs])
+
   return (
     <>
       <HeaderSlotPortal>
@@ -337,8 +207,6 @@ export function GeneConvPage() {
       <ShortcutLayout signinRequired={false}>
         <Toolbar>
           <ToolbarMenu
-            groupId={_id}
-            tabs={tabs}
             open={showFileMenu}
             onOpenChange={setShowFileMenu}
             fileMenuTabs={fileMenuTabs}
@@ -350,8 +218,6 @@ export function GeneConvPage() {
             }
           />
           <ToolbarPanel
-            groupId={_id}
-            tabs={tabs}
             tabShortcutMenu={
               <ShowOptionsMenu
                 show={showSideBar}