+++ src/components/pages/apps/genomic/get-dna/get-dna-page.tsx
@@ -352,11 +352,9 @@ export function GetDNAPage() {
                   <ToggleButtons
                     tabs={[
                       {
-                        //name: nanoid(),
                         id: 'FASTA',
                       },
                       {
-                        //name: nanoid(),
                         id: 'JSON',
                       },
                     ]}